backend/requirements/common.txt
asgiref==3.6.0
boto3==1.26.84
botocore==1.29.84
certifi==2022.12.7
charset-normalizer==3.0.1
Django==3.2.18
django-cors-headers==3.14.0
django-debug-toolbar==3.8.1
django-filter==22.1
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
mysqlclient==2.1.1
Pillow==9.4.0
PyJWT==2.6.0
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14